Error de espacios arreglado, controlador no funcion el buscar puestos desocupados

+++ Models/espacios.cs
@@ -8,7 +8,7 @@ namespace P01_2022_SS_650_2021_OF_601.Models
         public int id_espacios { get; set; }
         public int numero { get; set; }
         public string ubicacion { get; set; }
-        public float costo_hora { get; set; }
+        public double costo_hora { get; set; }
         public int id_sucursal { get; set; }
     }
 }
